requirements.txt
# Core Libraries
ultralytics>=8.0.0,<9.0.0
opencv-python>=4.5.0,<5.0.0
numpy>=1.21.0,<1.25.0
albumentations>=1.3.0,<1.6.0
matplotlib>=3.5.0,<4.0.0
pillow>=8.0.0,<10.0.0
pandas>=1.3.0,<2.0.0
tqdm>=4.65.0,<5.0.0
pyyaml>=6.0,<7.0.0


pytest>=7.0.0,<8.0.0


scikit-image>=0.19.0,<0.24.0
networkx>=3.0,<4.0.0

